+++ clustergenerator/src/main/java/clusterproject/clustergenerator/userInterface/MetaClustering/IDistanceMeasure.java
@@ -3,5 +3,5 @@ package clusterproject.clustergenerator.userInterface.MetaClustering;
 import clusterproject.clustergenerator.data.ClusteringResult;
 
 public interface IDistanceMeasure {
-	public float distanceBetween(ClusteringResult clustering, ClusteringResult clustering2);
+	public double distanceBetween(ClusteringResult clustering, ClusteringResult clustering2);
 }
